% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2grattex.R
\name{bib2grattex}
\alias{bib2grattex}
\title{Convert Word in-text citations to LaTeX \Cref citations using a .bib file.}
\usage{
bib2grattex(path = ".", bibName, texName, fromWord2grattex = FALSE,
  noBrackets = FALSE, titleLength = 20,
  variousYears = c("multiple years", "Various years", "various years"),
  ibid = TRUE, testRun = FALSE)
}
\arguments{
\item{path}{The path of the folder containing your Word document.}

\item{bibName}{The name of the .bib file containing all bib entries.}

\item{texName}{The name fo the .tex file containing in-text citations to be replaced.}

\item{fromWord2grattex}{TRUE if the bib2grattex function is nested within word2grattex.}

\item{titleLength}{The number of characters of the document title to be used in the reference key. The reference key is formatted as: AuthorYearTitle. Eg. Norton2018Droppingoutthecostsa when titleLength = 20 (the default).}

\item{variousYears}{Phrases that have been used to denote various years (in place of a numerical year within the in-text citation).}

\item{ibid}{Search for instances of 'ibid' and replace citation with most recent citation. This _only_ works when "ibid" is in the same paragraph as the most recent citation.}

\item{testRun}{To test new features of word2grattex. Leave as FALSE.}
}
\description{
Convert Word in-text citations to LaTeX \Cref citations using a .bib file.
}
\examples{
\dontrun{
library(word2grattex)
bib2grattex(path = path/to/worddoc/folder, bibName = "mybib.bib", texName = "myReport.tex")
}
}
