% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_figure_environments.R
\name{build_figure_environments}
\alias{build_figure_environments}
\title{Build table _environments_ for cross-references. But this does not build tables :(}
\usage{
build_figure_environments(tex_file, haveDownloadedGrattex = FALSE)
}
\arguments{
\item{tex_file}{The path to the .tex file}

\item{haveDownloadedGrattex}{logical: is this part of word2grattex AND are we downloading grattex?}
}
\description{
Build table _environments_ for cross-references. But this does not build tables :(
}
