% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_table_environments.R
\name{build_table_environments}
\alias{build_table_environments}
\title{Build table _environments_ for cross-references. But this does not build tables :(}
\usage{
build_table_environments(tex_file)
}
\arguments{
\item{tex_file}{The path to the .tex file}
}
\description{
Build table _environments_ for cross-references. But this does not build tables :(
}
