% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bib_duplicates.R
\name{check_bib_duplicates}
\alias{check_bib_duplicates}
\title{Check bib duplicates}
\usage{
check_bib_duplicates(bibPath, authorDistance = 0.4,
  titleDistance = 0.2, matchMethod = "jw", skipTitles = "")
}
\arguments{
\item{bibPath}{Path to the bib file}

\item{authorDistance}{A number between 0 and 1 for the loose (0) or tighter (1) match on authors (default is 0.4)}

\item{titleDistance}{A number between 0 and 1 for the loose (0) or tighter (1) match on titles (default is 0.2)}

\item{matchMethod}{"jw" is the default. See https://www.rdocumentation.org/packages/stringdist/versions/0.9.5.1/topics/stringdist-metrics for extensive options.}

\item{skipTitles}{A vector of bib entry titles to ignore. For example, "Budget Papers 1" will incorrectly match with "Budget Papers 2".
Add skipTitles = c("Budget Papers 1") to ignore this duplicate.}
}
\description{
Check bib duplicates
}
