% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up_pandoc.R
\name{clean_up_pandoc}
\alias{clean_up_pandoc}
\title{Clean up the mess created by pandoc during the conversion process.}
\usage{
clean_up_pandoc(tex_file)
}
\arguments{
\item{tex_file}{The path to the .tex file}
}
\description{
This function is an _art_ rather than a _science_.
}
