% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_doc_to_tex.R
\name{convert_doc_to_tex}
\alias{convert_doc_to_tex}
\title{Use pandoc to convert a Word document to a .tex file.}
\usage{
convert_doc_to_tex(file.docx, tex_file)
}
\arguments{
\item{file.docx}{The path to the .docx file to be converted.}

\item{tex_file}{The desired path of the exported .tex file}
}
\description{
Use pandoc to convert a Word document to a .tex file.
}
