% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_crossreferences.R
\name{replace_crossreferences}
\alias{replace_crossreferences}
\title{Replace Word and hard-coded cross-references with LaTeX `\\Cref`s}
\usage{
replace_crossreferences(tex_file)
}
\arguments{
\item{tex_file}{The path to the .tex file}
}
\description{
Replace Word and hard-coded cross-references with LaTeX `\\Cref`s
}
