% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_testing_ground.R
\name{run_testing_ground}
\alias{run_testing_ground}
\title{Run the 'testing ground' functions in word2grattex}
\usage{
run_testing_ground(tex_file)
}
\arguments{
\item{tex_file}{The path to the .tex file}
}
\description{
Run the 'testing ground' functions in word2grattex
}
