% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2grattex.R
\name{word2grattex}
\alias{word2grattex}
\title{Convert a Grattan report-style Word document to .tex}
\usage{
word2grattex(
  path = ".",
  overwrite = TRUE,
  downloadGrattex = FALSE,
  removeReport.tex = TRUE,
  bibReplace = TRUE,
  buildFigures = TRUE,
  buildTables = TRUE,
  crossReferences = TRUE,
  footnoteNewLine = TRUE,
  segmented = FALSE,
  testRun = FALSE
)
}
\arguments{
\item{path}{The path of the folder containing your Word document.}

\item{overwrite}{Overwrite an existing .tex file.}

\item{downloadGrattex}{Download the most recent version of Grattex. Set to TRUE if the Word document is not located within a Grattex folder.}

\item{removeReport.tex}{Delete the default Grattex .tex file.}

\item{bibReplace}{Add in-text citations using bib2grattex.}

\item{buildFigures}{Automatically build figure environments with titles, units, \\includegraphics, notes and sources.}

\item{buildTables}{Automatically build table environments to enable table cross-referencing. Note that this does _not_ build the table itself.}

\item{crossReferences}{Automatically create cross-references. Eg: “See Section 2.2” to "See \\Cref{subsec:section-name}".}

\item{footnoteNewLine}{Create a new line for each footnote in .tex}

\item{segmented}{Set to TRUE if this is just a chapter or section of a larger document.}

\item{testRun}{To test new features of word2grattex. Leave as FALSE.}
}
\description{
Convert a Grattan report-style Word document to .tex
}
