% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_footnote_newline.R
\name{add_footnote_newline}
\alias{add_footnote_newline}
\title{Create a new line for each footnote .tex}
\usage{
add_footnote_newline(texPath)
}
\arguments{
\item{texPath}{The path to a tex file, with the .tex extension.
Eg "report/myreport.tex"}
}
\description{
Create a new line for each footnote .tex
}
